Cached helper is moved down to platform dir
